- finish RGB<->GRAY conversions in convert.c - fix some (no where near
- finish RGB<->GRAY conversions in convert.c
- fix some (no where near all) -Wall -ansi -pedantic warnings
- drawable_type () is now gone
- replace valid_combinations array in gimage.c with a routine
  that examines and validates src/dest tags
- use drawable_width (), drawable_height () instead of directly
  accessing drawable member variables
- eliminate drawable member vars width, height, tag, delegate to
  tiles object

ray lehtiniemi <rayl@netrover.com>

+++ app/transform_tool.c
@@ -186,7 +186,7 @@ tools_free_transform_tool (Tool *tool)
 static void
 transform_change_type (int new_type)
 {
-  if (transform_options->type != new_type)
+  if ((int) transform_options->type != new_type)
     {
       /*  change the type, free the old tool, create the new tool  */
       transform_options->type = new_type;
